gcc warnings

+++ prboom2/src/gl_main.c
@@ -344,7 +344,7 @@ void gld_Init(int width, int height)
         p++;
       if (*p)
       {
-        int len = MIN(p-rover, sizeof(ext_name)-1);
+        size_t len = MIN(p-rover, sizeof(ext_name)-1);
         memset(ext_name, 0, sizeof(ext_name));
         strncpy(ext_name, rover, len);
         lprintf(LO_INFO,"\t%s\n", ext_name);
